+++ cookiecutter/hooks/post_gen_project.py
@@ -8,7 +8,11 @@ PROJECT_DIRECTORY = os.path.realpath(os.path.curdir)
 
 
 def remove_file(filepath):
-    os.remove(str(Path(PROJECT_DIRECTORY) / filepath))
+    f = Path(PROJECT_DIRECTORY) / filepath
+    if f.is_dir():
+        shutil.rmtree(str(f))
+    else:
+        f.unlink()
 
 
 def replace_contents(filename, what, replacement):
@@ -58,6 +62,7 @@ if __name__ == "__main__":
                    ("py34", "{{ cookiecutter.py34 }}"),
                    ("py35", "{{ cookiecutter.py35 }}"),
                    ("py36", "{{ cookiecutter.py36 }}"),
+                   ("py37", "{{ cookiecutter.py37 }}"),
                   ]:
         if on == "yes":
             tox_envlist.append(py)
@@ -95,3 +100,9 @@ if __name__ == "__main__":
         gk = Path("locale/.gitkeep")
         if gk.exists():
             gk.unlink()
+
+    if '{{ cookiecutter.add_docs }}' == 'no':
+        remove_file('docs')
+
+    if '{{ cookiecutter.requirements_yaml }}' == 'no':
+        remove_file('requirements')