+++ nicfit/__main__.py
@@ -28,7 +28,7 @@ MERGE_TOOLS["gvimdiff"] = "-geometry 169x60 -f"
 MERGE_TOOLS["vimdiff"] = None
 
 
-@nicfit.command.register
+@nicfit.Command.register
 class Requirements(nicfit.Command):
     """
     TODO
@@ -100,7 +100,7 @@ class Requirements(nicfit.Command):
         self._makeReqsFile(Path("requirements.txt"), pkg_reqs)
 
 
-@nicfit.command.register
+@nicfit.Command.register
 class CookieCutter(nicfit.Command):
     NAME = "cookiecutter"
     HELP = "Create a nicfit.py Python project skeleton."
@@ -302,10 +302,9 @@ class Nicfit(nicfit.Application):
     def __init__(self):
         super().__init__(version=version, gettext_domain="nicfit.py",
                          pdb_opt=True)
-        subs = self.arg_parser.add_subparsers(title="Commands",
-                                              add_help_subcmd=True,
-                                              dest="command")
-        nicfit.Command.initAll(subs)
+        subparsers = self.arg_parser.add_subparsers(dest="command",
+                                                    required=False)
+        nicfit.Command.loadCommandMap(subparsers=subparsers)
 
     def _main(self, args):
         ansi.init()