+++ nicfit/_argparse.py
@@ -1,4 +1,7 @@
+import sys
 import argparse
+import gettext
+_ = gettext.gettext
 
 
 class ArgumentParser(argparse.ArgumentParser):
@@ -13,15 +16,24 @@ class ArgumentParser(argparse.ArgumentParser):
         self._add_log_args = add_log_args
 
         if config_opts:
-            from . import _config
-            _config.addCommandLineArgs(self, config_opts)
+            from . import config
+            config.addCommandLineArgs(self, config_opts)
         self._config_opts = config_opts
+        # For python <= 3.6, where subcmds are optional
+        self._subcmd_required = None
 
     def parse_known_args(self, args=None, namespace=None):
         from . import logger
 
         parsed, remaining = super().parse_known_args(args=args,
                                                      namespace=namespace)
+
+        # Required sub command support for Python < 3.7
+        if self._subcmd_required is not None:
+            req, dest = self._subcmd_required
+            if req and dest and (not hasattr(parsed, dest) or
+                                 not getattr(parsed, dest)):
+                self.error(_('the following arguments are required: %s') % dest)
 
         if "config" in parsed and "config_overrides" in parsed:
             config = parsed.config
@@ -43,16 +55,23 @@ class ArgumentParser(argparse.ArgumentParser):
 
         return parsed, remaining
 
-    def add_subparsers(self, add_help_subcmd=False, **kwargs):
+    def add_subparsers(self, add_help_subcmd=False, required=True,
+                       dest="subcmd", **kwargs):
         if "parser_class" not in kwargs:
             kwargs["parser_class"] = ArgumentParser
 
-        subparser = super().add_subparsers(**kwargs)
+        if sys.version_info[:2] >= (3, 7):
+            subparser = super().add_subparsers(required=required, **kwargs)
+        else:
+            self._subcmd_required = (
+                required, dest,
+            )
+            subparser = super().add_subparsers(dest=dest, **kwargs)
 
         if add_help_subcmd:
             # 'help' subcommand; turns it into the less intuitive --help format.
             # e.g.  cmd help subcmd  ==> cmd subcmd --help
-            def _help(args, config):
+            def _help(args):
                 if args.command:
                     self.parse_args([args.command, "--help"])
                 else:
@@ -61,7 +80,7 @@ class ArgumentParser(argparse.ArgumentParser):
 
             help = subparser.add_parser("help",
                                         help="Show help for a sub command")
-            help.set_defaults(func=_help)
+            help.set_defaults(command_func=_help)
             help.add_argument("command", nargs='?', default=None)
 
         return subparser