+++ tests/test_logging.py
@@ -105,25 +105,6 @@ def test_logFileArguments(tmpdir):
         # FIXME: actually test
 
 
-@deprecation.fail_if_not_removed
-def test_FileConfig_OLD():
-    from nicfit.logger import LOGGING_CONFIG
-    cfg = LOGGING_CONFIG("INl3agueWitS4t4n")
-    cfg_file = StringIO(cfg)
-    logging.config.fileConfig(cfg_file)
-
-
-@deprecation.fail_if_not_removed
-def test_FileConfig_auto_OLD():
-    from nicfit.logger import LOGGING_CONFIG
-    with patch("logging.config.fileConfig") as mock_fileConfig:
-        LOGGING_CONFIG("INl3agueWitS4t4n", init_logging=True)
-        if sys.version_info[:2] >= (3, 6):
-            mock_fileConfig.assert_called()
-        else:
-            assert mock_fileConfig.call_count != 0
-
-
 def test_FileConfig():
     cfg = "\n".join([FileConfig.DEFAULT_LOGGING_CONFIG(),
                      FileConfig.PKG_LOGGING_CONFIG("INl3agueWitS4t4n")])